+++ Server/routes/admins.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const {registerAdmins, findAll, deleteById, login } = require('../controller/admins'); 
+const {registerAdmins, findAll, deleteById, login, searchAdmins } = require('../controller/admins'); 
 
 const router = express.Router();
 
@@ -7,6 +7,7 @@ router.get('/', findAll);
 router.post('/register', registerAdmins);
 router.post('/login', login)
 router.delete('/delete', deleteById);
+router.get('/:name', searchAdmins)
 
 module.exports = router;
